requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
boto3
pandas
aiohttp
bcrypt
python-multipart
dnspython
